[tlsconfig] include the HTTP/2 required suite
We can't use the recommended curve so maybe this is all for nothing? The
HTTP/2 spec requires use of P-256 which the DoD explicitly does not
allow. Since this is for internal services only where we control both
ends of the connection I don't think this will affect us but caution
should be practiced when attempting to connect to outside services.

[1]: https://http2.github.io/http2-spec/#rfc.section.9.2.2