upgrade rules_go version
This version could not gazelle the grpc_test package from grpc-go so
I've written a similar one in test/grpctest. This is probably a good
idea anyway to avoid depending on their package.

The certstrap package has been changed back to upstream after Bazel will
ignore the "build" script in their root directory and instead use
"BUILD.bazel" by default. They changed an API which is also in this
commit. There should be no changes for users of this package unless they
were depending on the expiration time of the certificates.